Check whether ColorView is defined (#67)
ColorView has been deprecated since Nov 2017, but only as a constructor.
Hence direct uses of the type did not get noticed. I may "resurrect"
`ColorView` in new form, and if so the existing definition here should still
work, but in case we go through a period where there is no `ColorView`
we should check for its existence.